Add DockerContainerImage#LoadImage
This method allows Docker Images to be imported from any
io.ReadCloser. It also accepts a name parameter, which denotes the
expected name of the imported image. Using the name allows the method to
determine whether the image is already in Docker, which is a convenient
short circuit. The method also ensures that the expected name was loaded
after the operation, raising an error if it wasn't.